# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * India Election Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalElections: number;
  liveElections: number;
  upcomingElections: number;
  completedElections: number;
  totalCandidates: number;
  totalVotesCast: number;
}