# lib/api-zod/src/generated/types/createElectionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * India Election Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateElectionRequestElectionType } from "./createElectionRequestElectionType";
import type { CreateElectionRequestStatus } from "./createElectionRequestStatus";

export interface CreateElectionRequest {
  title: string;
  description?: string;
  electionType: CreateElectionRequestElectionType;
  status: CreateElectionRequestStatus;
  startDate: string;
  endDate: string;
  state: string;
  constituency?: string;
}