# lib/api-zod/src/generated/types/electionWithCandidates.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * India Election Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from "./candidate";
import type { ElectionWithCandidatesElectionType } from "./electionWithCandidatesElectionType";
import type { ElectionWithCandidatesStatus } from "./electionWithCandidatesStatus";

export interface ElectionWithCandidates {
  id: number;
  title: string;
  description?: string;
  electionType: ElectionWithCandidatesElectionType;
  status: ElectionWithCandidatesStatus;
  startDate: string;
  endDate: string;
  state: string;
  constituency?: string;
  createdAt: string;
  candidates: Candidate[];
}